LayoutTests/imported/w3c/web-platform-tests/svg/text/scripted/getrotationofchar-expected.txt

M月X
T火
W水
t木cde
F金r
M月X
T火
W水
t木cde
F金r
M月X
T火
W水
t木cde
F金r
M月X
T火
W水
t木cde
F金r
M月X
T火
W水
t木cde
F金r

PASS 'text-orientation: mixed' - horizontal-tb
PASS 'text-orientation: upright' - horizontal-tb
PASS 'text-orientation: sideways' - horizontal-tb
FAIL Rotation attribute - horizontal-tb assert_equals: c expected 0 but got 360
FAIL <textPath> - horizontal-tb assert_equals: r expected 70 but got 45
PASS 'text-orientation: mixed' - vertical-rl
FAIL 'text-orientation: upright' - vertical-rl assert_equals: T expected 0 but got 90
FAIL 'text-orientation: sideways' - vertical-rl assert_equals: 水 expected 90 but got 0
PASS Rotation attribute - vertical-rl
FAIL <textPath> - vertical-rl assert_equals: r expected 70 but got 45
PASS 'text-orientation: mixed' - vertical-lr
FAIL 'text-orientation: upright' - vertical-lr assert_equals: T expected 0 but got 90
FAIL 'text-orientation: sideways' - vertical-lr assert_equals: 水 expected 90 but got 0
PASS Rotation attribute - vertical-lr
FAIL <textPath> - vertical-lr assert_equals: r expected 70 but got 45
FAIL 'text-orientation: mixed' - sideways-rl assert_equals: 月 expected 90 but got 0
FAIL 'text-orientation: upright' - sideways-rl assert_equals: 火 expected 90 but got 0
FAIL 'text-orientation: sideways' - sideways-rl assert_equals: 水 expected 90 but got 0
FAIL Rotation attribute - sideways-rl assert_equals: 木 expected 270.5 but got 180.5
FAIL <textPath> - sideways-rl assert_equals: 金 expected 45 but got 315
FAIL 'text-orientation: mixed' - sideways-lr assert_equals: M expected -90 but got 90
FAIL 'text-orientation: upright' - sideways-lr assert_equals: T expected -90 but got 90
FAIL 'text-orientation: sideways' - sideways-lr assert_equals: W expected -90 but got 90
FAIL Rotation attribute - sideways-lr assert_equals: t expected 270 but got 90
FAIL <textPath> - sideways-lr assert_equals: F expected 225 but got 45
